Move auth.js to lib folder
