+++ client/lib/auth.js
